adding follower and camera holder components/systems
Camera Holder: Configures the internal camera from Camera components to
always target the global position of it's self entity

Follower: Follows another position, however, be cautious:

it runs before the position hierarchy system, so it only will work
properly if the followed position is from a sibling entity.

Only use this when making the "follower" a child of the "followed" is
not possible.

lunar_lander.nelua will be committed later