add scene graph node component and position hierarchy system
Scene graphs are used to make a tree (or forests) of entities, here, the
SceneGraphNode component holds the information about itself, the parent,
and the children nodes.

The implementation of a position hierarchy is on the PositionHierarchy
system.

It's also possible to create others hierarchies systems, like an
hierarchy of AABB of intersectables which the parent encapsulates the
children AABB intersectables. (probably, not tested)