Add Lander Propulsion entity as a child of Lander entity
I discovered something REALLY important:

ALWAYS instantiates the children entities first, more especifically,
before getting any component.

The reason is, when you get a component, you're getting a pointer of
this component which "lives" on the `entries` vector field of the
storage, however, when a a new component is created, this `entries`
field may be reallocated, and then your pointer will be invalidated.

And it's a little hard to debug your code to find this problem.

Because of that, you should first create the entities, and then with
everything setup, manipulate the components.