add more "get_.." methods
Every "Entity Type" now have an `get_entity` method, it will fail if the
entity could not be found

There is also a new "get_" methods for each child, similar to the
components ones.

It uses the `get_entity` method, so it will fail if the entity could not
be found.