WIP: SDL2 refactor
For now I'm creating a basic micro (nano?) game framework (quickgamelib)
and game math library (gamemath).

The "intersection test" and "basic gravity" are also refactored to use
quickgamelib.

I still want to use tinycoffee, but is still a WIP library while SDL2 is
battle tested, which is something I need to make a new game soon.

I may make "quickgamelib" compatible with both SDL2 and tinycoffee in
the future, so the transition between the two libraries will be easier.