makers make_entity refactor
The main changes are about what fields an "entity shape" (EntityT)
contains; the component/child getter functions; "entity
shapes" are now also components; and new functions for easily destroying
an entity associated with the "shape".

An "entity shape" (provisory name for the type returned by MakeEntity)
is also a component, and an entity made by this "entity shape" contains
that "shape" as a component automatically.

That way, if you need this shape from an entity_id, you can access
through
`RotorQuick.storages.entity_name.get_entry_from_entity_id(entity_id)`.

The entity also now contains the components of the children "shapes"
(but this is untested).

There is now two new functions for easily destroying the entity and it's
components: `.destroy` and `.destroy_from_entity_id`. Is possible to
pass callbacks for each component associated with the entity, so the
components can be properly destroyed.

Finally, a very important breaking change: the `:get_"component/child"`
methods are removed and replaced with `:get_component` and `:get_child`
methods, you should pass the type of the component in order to get it,
the same applies for children.

Signed-off-by: André Luiz Alvares <andrluiz.alvares.alvares84@gmail.com>