Introduce `map_caller` and `arrmap`, `textmap` callbacks moved to `map_caller`
Instead of giving a text map and the callbacks to `textmap`,
now `textmap` only receives the text and returns a map matrix (in other
words, it's just a conversion).

The callback is now done at `map_caller`, which receives this matrix.

Along with that, now there is a `arrmap`, which receives an array, which
then it's converted to a matrix (this is useful on Tiled Lua export).