Formalize some repeated logic on a "Painter" component and add layers.
Layers are just `byte` numbers, and 0 by default thanks to ZII of nelua.

There's no example that uses it so I'll explain in this commit message:

Just set the `active_layer` with the "active painting layer" of the
running system, thus entities with an different layer will not be painted.

This means that you can run the painting systems on a `for` loop, thus,
getting a layering system without using sorting.

As always, this have advantages (it's simpler than making an actual
system), but also disadvantages (running the same systems multiple times).